@@ -14,20 +14,21 @@ if (localPropertiesFile.exists()) {
 
 android {
     namespace = "com.example.myapplication"
-    compileSdk = 34
+    compileSdk = 35
 
     buildFeatures {
         viewBinding = true
+        buildConfig = true
     }
 
     defaultConfig {
         applicationId = "com.example.myapplication"
         minSdk = 30
-        targetSdk = 34
+        targetSdk = 35
         versionCode = 1
         versionName = "1.0"
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        buildConfigField("String", "KAKAO_MAP_KEY", properties.getProperty("KAKAO_MAP_KEY"))
+//        buildConfigField("String", "KAKAO_MAP_KEY", properties.getProperty("KAKAO_MAP_KEY"))
     }
 
     buildTypes {
@@ -50,51 +51,57 @@ android {
 
 dependencies {
     // Android 기본 라이브러리
-    implementation("androidx.core:core-ktx:1.10.1")
-    implementation("androidx.appcompat:appcompat:1.6.1")
-    implementation("com.google.android.material:material:1.9.0")
-    implementation("androidx.activity:activity-ktx:1.7.2")
-    implementation("androidx.constraintlayout:constraintlayout:2.1.4")
-    implementation("com.google.android.gms:play-services-maps:18.1.0")
+    implementation(libs.androidx.core.ktx)
+    implementation("androidx.appcompat:appcompat:1.7.0")
+    implementation(libs.material)
+    implementation(libs.androidx.activity.ktx)
+    implementation(libs.androidx.constraintlayout)
+    implementation(libs.play.services.maps)
 
     // Glide (이미지 로딩)
-    implementation("com.github.bumptech.glide:glide:4.15.1")
+    implementation(libs.glide)
     implementation(libs.play.services.location)
-    annotationProcessor("com.github.bumptech.glide:compiler:4.15.1")
+    annotationProcessor(libs.compiler)
 
     // Retrofit (HTTP 요청)
-    implementation("com.squareup.retrofit2:retrofit:2.9.0")
-    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
-    implementation("com.squareup.retrofit2:converter-scalars:2.9.0")
+    implementation(libs.retrofit)
+    implementation(libs.converter.gson)
+    implementation(libs.converter.scalars)
 
-
-    // OkHttp (HTTP 클라이언트)
-    implementation("com.squareup.okhttp3:okhttp:4.9.3")
 
     // Gson (JSON 파싱)
-    implementation("com.google.code.gson:gson:2.9.0")
+    implementation(libs.gson)
 
     // 테스트 라이브러리
-    testImplementation("junit:junit:4.13.2")
-    androidTestImplementation("androidx.test.ext:junit:1.1.5")
-    androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
+    testImplementation(libs.junit)
+    androidTestImplementation(libs.androidx.junit.v115)
+    androidTestImplementation(libs.androidx.espresso.core)
 
     // 네트워크 라이브러리
-    implementation("com.squareup.okhttp3:okhttp:4.10.0")
+    implementation(libs.okhttp)
 
 
     // fragment
-    implementation ("androidx.fragment:fragment-ktx:1.6.1")
-    implementation ("androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.1")
-    implementation ("androidx.lifecycle:lifecycle-livedata-ktx:2.6.1")
+    implementation (libs.androidx.fragment.ktx)
+    implementation (libs.androidx.lifecycle.viewmodel.ktx)
+    implementation (libs.androidx.lifecycle.livedata.ktx)
 
 
-    // Mapbox
-    implementation ("com.mapbox.mapboxsdk:mapbox-android-sdk:10.13.0")
+//    // Mapbox
+//    implementation ("com.mapbox.mapboxsdk:mapbox-android-sdk:11.8.1")
 
     implementation(libs.play.services.maps) // Google Maps
+    implementation ("com.google.android.gms:play-services-location:21.0.1")
+
     implementation(libs.naver.map.sdk) // Naver Maps
-    implementation(libs.kakao.sdk) // Kakao API
+//    implementation(libs.kakao.sdk) // Kakao API
+
+//    // Speech-to-Text API
+//    implementation ("com.google.cloud:google-cloud-speech:2.3.0")
+//
+//    // Text-to-Speech API
+//    implementation ("com.google.cloud:google-cloud-texttospeech:2.3.0")
+
 
 }
 
